+++ Options/CHANGELOG.lua
@@ -4,6 +4,7 @@ TMW.CHANGELOG_LASTVER="10.0.0"
 
 TMW.CHANGELOG = [==[
 ## v11.2.7
+* Fix #2309 Additional assorted talent/spec errors on Classic Era.
 
 ## v11.2.6
 * Fix #2307 C_SpecializationInfo.GetTalent: query.specializationIndex must be specified.