+++ TellMeWhen.lua
@@ -175,8 +175,8 @@ end)
 
 ---------- Upvalues ----------
 local GetSpellTexture = C_Spell and C_Spell.GetSpellTexture or GetSpellTexture
-local InCombatLockdown, GetTalentInfo =
-	  InCombatLockdown, GetTalentInfo
+local InCombatLockdown =
+	  InCombatLockdown
 local IsInGuild, IsInGroup, IsInInstance =
 	  IsInGuild, IsInGroup, IsInInstance
 local tonumber, tostring, type, pairs, ipairs, tinsert, tremove, sort, select, wipe, rawget, rawset, assert, pcall, error, getmetatable, setmetatable, loadstring, unpack, debugstack =
@@ -1083,9 +1083,7 @@ function TMW:PLAYER_LOGIN()
 	-- end
 	
 
-
-	TMW:UpdateTalentTextureCache()
-
+	safecall(TMW.UpdateTalentTextureCache, TMW)
 
 	TMW:RegisterEvent("GLOBAL_MOUSE_DOWN")
 	
@@ -2829,51 +2827,16 @@ function TMW:ScheduleUpdate(delay)
 end
 
 function TMW:UpdateTalentTextureCache()
-	if C_SpecializationInfo.GetTalentInfo and MAX_NUM_TALENT_TIERS and NUM_TALENT_COLUMNS then
-		-- Should handle all classic versions mop and below?
-		local talentInfoQuery = {};
+	for _, talentInfoQuery in TMW.GetTalentQueries() do
+		local talentInfo = C_SpecializationInfo.GetTalentInfo(talentInfoQuery);
+		if talentInfo then
+			local name = talentInfo.name
+			local tex = talentInfo.fileID
 
-		for spec = 1, TMW.GetNumSpecializations() do
-			for tier = 1, MAX_NUM_TALENT_TIERS do
-				for column = 1, NUM_TALENT_COLUMNS do
-					talentInfoQuery.tier = tier;
-					talentInfoQuery.column = column;
-					talentInfoQuery.specializationIndex = spec;
-					local talentInfo = C_SpecializationInfo.GetTalentInfo(talentInfoQuery);
+			local lower = name and strlowerCache[name]
 
-					local name = talentInfo.name
-					local tex = talentInfo.fileID
-
-					local lower = name and strlowerCache[name]
-
-					if lower then
-						SpellTexturesMetaIndex[lower] = tex
-					end
-				end
-			end
-		end
-	elseif MAX_TALENT_TIERS and NUM_TALENT_COLUMNS and GetTalentInfo then
-		for tier = 1, MAX_TALENT_TIERS do
-			for column = 1, NUM_TALENT_COLUMNS do
-				local id, name, tex = GetTalentInfo(tier, column, 1)
-
-				local lower = name and strlowerCache[name]
-				
-				if lower then
-					SpellTexturesMetaIndex[lower] = tex
-				end
-			end
-		end
-	elseif GetNumTalentTabs then
-		for tab = 1, GetNumTalentTabs() do
-			for index = 1, GetNumTalents(tab) do
-				local name, iconTexture = GetTalentInfo(tab, index)
-
-				local lower = name and strlowerCache[name]
-				
-				if lower then
-					SpellTexturesMetaIndex[lower] = iconTexture
-				end
+			if lower then
+				SpellTexturesMetaIndex[lower] = tex
 			end
 		end
 	end