+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v11.2.7
+* Fix #2309 Additional assorted talent/spec errors on Classic Era.
 
 ## v11.2.6
 * Fix #2307 C_SpecializationInfo.GetTalent: query.specializationIndex must be specified.